Round MinimumFee, AdminChargerDto.PriceValue and remaining money fields to int
Close out the response-DTO money-rounding rollout: convert
FeesConfigDto.MinimumFee, AdminFeesDto.MinimumFee,
AdminChargerDto.PriceValue and AdminReservationDto.PaidAmount to int via
MoneyRounding at the mapping/service boundary. Percentage stays decimal
(factor, supports 10.5%), and the Wallet surface (entity, balances,
transaction history) is still out of scope pending a dedicated refactor.

+++ Voltyks.AdminControlDashboard/Dtos/Chargers/AdminChargerDto.cs
@@ -15,7 +15,7 @@ namespace Voltyks.AdminControlDashboard.Dtos.Chargers
         public int CapacityKw { get; set; }
 
         public int PriceOptionId { get; set; }
-        public decimal PriceValue { get; set; }
+        public int PriceValue { get; set; }
 
         public int AddressId { get; set; }
         public string Area { get; set; } = "";
