+++ Voltyks.AdminControlDashboard/Dtos/Fees/AdminFeesDto.cs
@@ -3,7 +3,7 @@ namespace Voltyks.AdminControlDashboard.Dtos.Fees
     public class AdminFeesDto
     {
         public int Id { get; set; }
-        public decimal MinimumFee { get; set; }
+        public int MinimumFee { get; set; }
         public decimal Percentage { get; set; }
         public DateTime UpdatedAt { get; set; }
         public string? UpdatedBy { get; set; }
